+++ misc/python/materialize/xcompile.py
@@ -111,6 +111,7 @@ def cargo(
 
     env = {
         **extra_env,
+        "CARGO_TARGET_DIR": ROOT / "target-xcompile",
         "RUSTFLAGS": " ".join(rustflags),
         **KRB5_CONF_OVERRIDES,
     }
@@ -160,7 +161,7 @@ def _bootstrap_darwin(arch: Arch) -> None:
     # cross-compiling toolchain on the host and use that instead.
 
     BOOTSTRAP_VERSION = "4"
-    BOOTSTRAP_FILE = ROOT / "target" / target(arch) / ".xcompile-bootstrap"
+    BOOTSTRAP_FILE = ROOT / "target-xcompile" / target(arch) / ".xcompile-bootstrap"
     try:
         contents = BOOTSTRAP_FILE.read_text()
     except FileNotFoundError:
@@ -169,7 +170,6 @@ def _bootstrap_darwin(arch: Arch) -> None:
         return
 
     spawn.runv(["brew", "install", f"materializeinc/crosstools/{target(arch)}"])
-    spawn.runv(["cargo", "clean", "--target", target(arch)], cwd=ROOT)
     spawn.runv(["rustup", "target", "add", target(arch)])
 
     BOOTSTRAP_FILE.parent.mkdir(parents=True, exist_ok=True)