+++ src/coord/src/catalog/config.rs
@@ -10,6 +10,7 @@
 use std::time::Duration;
 
 use mz_build_info::BuildInfo;
+use mz_dataflow_types::sources::AwsExternalId;
 use mz_ore::metrics::MetricsRegistry;
 
 use crate::catalog::storage;
@@ -31,7 +32,7 @@ pub struct Config<'a> {
     /// An [External ID][] to use for all AWS AssumeRole operations.
     ///
     /// [External ID]: https://docs.aws.amazon.com/IAM/latest/UserGuide/id_roles_create_for-user_externalid.html
-    pub aws_external_id: Option<String>,
+    pub aws_external_id: AwsExternalId,
     /// Timestamp frequency to use for CREATE SOURCE
     pub timestamp_frequency: Duration,
     /// Function to generate wall clock now; can be mocked.