+++ src/ore/src/lib.rs
@@ -63,8 +63,8 @@ pub mod retry;
 pub mod stack;
 pub mod stats;
 pub mod str;
+#[cfg_attr(nightly_doc_features, doc(cfg(feature = "task")))]
 #[cfg(feature = "task")]
-#[doc(hidden)]
 pub mod task;
 #[cfg_attr(nightly_doc_features, doc(cfg(feature = "test")))]
 #[cfg(feature = "test")]